+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for KiCad 5 "could not find valid layer for pad" warning
+  (See INTI-CMNB/KiBot#299)
 
 ## [2.0.4] - 2022-09-12
 ### Added